Add benchmark tests
Run PropertiesBenchmarkIT (either on command line or IDE). Make sure
to refresh the Maven project in Eclipse first if that's where you are
(to generate the benchmark classes with JMH).

Also set -DpublishTo=csv:target/result.csv to get the result printed
in CSV format (easier to read).

Run BinderAllocations as a main class using an Oracle JDK and set

    -XX:+UnlockCommercialFeatures -XX:+FlightRecorder \
    -XX:StartFlightRecording=delay=0s,duration=120s,name=recording,filename=/tmp/recording.jfr,settings=profile

to record profiling data.